- Create Makefile for easy compile and create a docker image - Create docker-compose for easy run environment locally - Prepare helm chart - Add golang code for implement monitoring 4 Golden Signal
