+++ devops/docker-compose.yml
@@ -1,0 +1,46 @@
+version: '3.8'
+services:
+  app:
+    image: app:latest
+    environment:
+      - REDIS_ADDR=redis:6379
+    ports:
+      - 8080:8080
+      - 8081:8081
+
+  cache:
+    image: dockerhub.timeweb.cloud/bitnami/redis:7.2.5
+    restart: always
+    ports:
+      - 6379:6379
+    command: redis-server --save 20 1 --loglevel warning --daemonize yes
+    volumes: 
+      - cache:/data
+
+  prometheus:
+    image: dockerhub.timeweb.cloud/prom/prometheus:latest
+    container_name: prometheus
+    ports:
+      - "9090:9090"
+    volumes:
+      - ./monitoring/prometheus.yml:/etc/prometheus/prometheus.yml
+
+  grafana:
+    image: dockerhub.timeweb.cloud/grafana/grafana:latest
+    container_name: grafana
+    ports:
+      - "3000:3000"
+    depends_on:
+      - prometheus
+    environment:
+      GF_SECURITY_ADMIN_USER: admin
+      GF_SECURITY_ADMIN_PASSWORD: poengairi6baiKaiv2Ce2yoo1
+    volumes:
+      - grafana-storage:/var/lib/grafana
+      - ${PWD}/monitoring/datasource:/etc/grafana/provisioning/datasources
+      - ${PWD}/monitoring/dashboards:/etc/grafana/provisioning/dashboards
+
+volumes:
+  cache:
+    driver: local
+  grafana-storage:
