--- go.mod
@@ -1,11 +1,0 @@
-module github.com/nuggien/devops_test
-
-go 1.16
-
-require (
-	github.com/go-redis/redis/v8 v8.11.2
-	github.com/julienschmidt/httprouter v1.3.0
-	github.com/lib/pq v1.10.2
-	github.com/rs/cors v1.8.0
-	go.uber.org/zap v1.19.0
-)
