+++ src/go.mod
@@ -1,0 +1,24 @@
+module github.com/nuggien/devops_test
+
+go 1.22
+
+require (
+	github.com/go-redis/redis/v8 v8.11.2
+	github.com/julienschmidt/httprouter v1.3.0
+	github.com/prometheus/client_golang v1.19.1
+	go.uber.org/zap v1.19.0
+
+)
+
+require (
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/prometheus/client_model v0.5.0 // indirect
+	github.com/prometheus/common v0.48.0 // indirect
+	github.com/prometheus/procfs v0.12.0 // indirect
+	go.uber.org/atomic v1.7.0 // indirect
+	go.uber.org/multierr v1.6.0 // indirect
+	golang.org/x/sys v0.17.0 // indirect
+	google.golang.org/protobuf v1.33.0 // indirect
+)
